asp_backend/asp_backend/requirements.txt
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
cryptography==41.0.7
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.4
pip==23.3.1
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.9.0
urllib==31.26.18
wheel==0.41.2